scripts/requirements.txt
flask==2.3.2
gunicorn==20.1.0
google-genai
python-dotenv
Pillow
requests
